pyproject.toml
[project]
name = "sim"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.40",
    "instructor>=1.3",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "jinja2>=3.1",
    "loguru>=0.7",
    "python-dotenv>=1.2.2",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "sse-starlette>=2.0",
    "pillow>=10.3",
]

[project.scripts]
sim = "sim.main:main"
api = "sim.api.server:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sim"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
